Moving the cellline parser to the shared folder.
